+++ backend/auth/auth-api/src/main/java/com/meemaw/auth/organization/datasource/OrganizationTeamInviteTable.java
@@ -1,6 +1,6 @@
 package com.meemaw.auth.organization.datasource;
 
-public final class OrganizationInviteTable {
+public final class OrganizationTeamInviteTable {
 
   public static final String TOKEN = "token";
   public static final String EMAIL = "email";
@@ -9,5 +9,5 @@ public final class OrganizationInviteTable {
   public static final String CREATOR_ID = "creator_id";
   public static final String CREATED_AT = "created_at";
 
-  private OrganizationInviteTable() {}
+  private OrganizationTeamInviteTable() {}
 }
